refactor: api로직 리팩터링
